Unified Diffs (#3)
* Initial cursed version of unified diffs...

I need to probably take a look at parsing and probalby defer a lot of
individual line parsing later on (both will save on performance and
probably be easier to switch between the types...)

Real goal would be -- is there a way i can use exactly the same HTML to
switch between the type? :thonk:

* Improved unified/split diff rendering

Refactored some aspects of diff parsing to make the render flows a bit
more compatible.  There's probably more I can clean this up but it's
nicer than it was in the prior commit

* Clean up some split vs unified code

* Rework addition/deletion styles to be more universal regardless of
  split vs unified
* Make css better organized/cleaner

* Adding another diff file, bcuz

* Make the unified diffs toggle usable after render

* Improve error logging

* Fix up some error logging